docs: Updated PowerShell comment
`Del` itself  is an alias and when working in a restricted PS environment or if the aliases have been altered already, `Del` could not work. Also added `-ErrorAction Ignore` to avoid failing if the `ni` alias doesn't exist.
The PS 7 command or both command sections could also be replaced by this, although the PS 7 command is more modern.